Split reporting.quantity and reporting.attrseries
- Make Quantity() a factory method rather than class/type.
- Quantity.CLASS is a string used to check the current implementation.
- Remove reporting.as_quantity and all uses of this function;
  Quantity(...) now does the same.
- Move AttrSeries.align_levels from computations.product.
- Update testing.assert_qty_equal and .assert_qty_allclose.
